Move RequestManager into core module & add kotlin-logging dependency
