+++ core/src/main/kotlin/org/jetbrains/research/testspark/core/data/ChatMessage.kt
@@ -1,0 +1,6 @@
+package org.jetbrains.research.testspark.core.data
+
+data class ChatMessage(
+    val role: String,
+    val content: String,
+)
