+++ src/main/kotlin/org/jetbrains/research/testspark/tools/llm/generation/IJRequestManager.kt
@@ -1,6 +1,7 @@
 package org.jetbrains.research.testspark.tools.llm.generation
 
 import com.intellij.openapi.project.Project
+import org.jetbrains.research.testspark.core.generation.llm.network.RequestManager
 import org.jetbrains.research.testspark.tools.llm.SettingsArguments
