+++ src/main/kotlin/org/jetbrains/research/testspark/tools/llm/generation/RequestManagerFactory.kt
@@ -1,6 +1,7 @@
 package org.jetbrains.research.testspark.tools.llm.generation
 
 import com.intellij.openapi.project.Project
+import org.jetbrains.research.testspark.core.generation.llm.network.RequestManager
 import org.jetbrains.research.testspark.services.SettingsApplicationService
 import org.jetbrains.research.testspark.settings.SettingsApplicationState
 import org.jetbrains.research.testspark.tools.llm.SettingsArguments
