+++ src/main/kotlin/org/jetbrains/research/testspark/tools/llm/generation/openai/OpenAIRequestBody.kt
@@ -1,12 +1,9 @@
 package org.jetbrains.research.testspark.tools.llm.generation.openai
 
+import org.jetbrains.research.testspark.core.data.ChatMessage
+
 data class OpenAIRequestBody(
     val model: String,
     val messages: List<ChatMessage>,
     val stream: Boolean = true,
-)
-
-data class ChatMessage(
-    val role: String,
-    val content: String,
 )
